Issue #18 Restoring tests for validating form uploaded config file.
